+++ src/index.ts
@@ -1,4 +1,5 @@
 export * as calendar from './lib/api/calendar';
 export * as youtube from './lib/api/youtube';
+export { getItems } from './lib/api/shared';
 export { createProfile, getProfiles } from './lib/profiles';
 export { login } from './lib/auth';