+++ src/lib/api/calendar.ts
@@ -4,8 +4,8 @@ import { getAuth } from '../auth';
 import { getItems } from './shared';
 import calendar from './calendar';
 
-export { getAPI, getEvents, setEvent };
-export default { getAPI, getEvents, setEvent };
+export { getAPI, getEvents, getCalendars, setEvent };
+export default { getAPI, getEvents, getCalendars, setEvent };
 
 async function getAPI(profile: string): Promise<GoogleApis.calendar_v3.Calendar> {
 	const googleAuth = await getAuth(profile);
@@ -15,6 +15,12 @@ async function getAPI(profile: string): Promise<GoogleApis.calendar_v3.Calendar>
 		auth    : googleAuth,
 	});
 }
+
+async function getCalendars(profile: string, args: GoogleApis.calendar_v3.Params$Resource$Calendarlist$List) {
+	const api = await calendar.getAPI(profile);
+	return getItems(api.calendarList, args);
+}
+
 async function getEvents(profile: string, args: GoogleApis.calendar_v3.Params$Resource$Events$List) {
 	const api = await calendar.getAPI(profile);
 	return getItems(api.events, args);