+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.2](../../tags/v2.0.2) - 2023-03-13
+### Added
+- Calendars list API
+- Export shared `getItems` to use for any other APIs
+
+## [2.0.1](../../tags/v2.0.1) - 2023-03-12
+### Changed
+- Fixed exported types
+
 ## [2.0.0](../../tags/v2.0.0) - 2023-03-12
 ### Changed
 - Split APIs